requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-http-client==3.3.7
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
sendgrid==6.11.0
sqlparse==0.5.2
starkbank-ecdsa==2.2.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
